backend/requirements.txt
pathway
google-generativeai
python-dotenv
psycopg2-binary
tomli
